HB2-40 split test files and refactor shit
