+++ homebytwo/templates/importers/routes.html
@@ -5,28 +5,34 @@
 
 {% block content %}
   <h1>Import Routes from {{ source.name }}</h1>
+
   {# New Routes#}
   {% if new_routes %}
     <h2>New Routes</h2>
+
     <ul class="routes list-stacked list-stacked--tight">
-    {% for route in new_routes %}
-      {% include "importers/_route_card.html" with new=True %}
-    {% endfor %}
+      {% for route in new_routes %}
+        {% include "importers/_route_card.html" with new=True %}
+      {% endfor %}
     </ul>
   {% endif %}
+
   {# Old Routes#}
   {% if old_routes %}
     <h2>Already Imported Routes</h2>
+
     <ul class="routes list-stacked list-stacked--tight">
-    {% for route in old_routes %}
-      {% include "importers/_route_card.html" %}
-    {% endfor %}
+      {% for route in old_routes %}
+        {% include "importers/_route_card.html" %}
+      {% endfor %}
+    </ul>
   {% endif %}
+
   {#No Routes#}
   {% if not new_routes and not old_routes %}
     <div class="box">
-    <h3>No route found</h3>
-    <p>Go create some at {{ source.name }} and come back here to import them.</p>
+      <h3>No route found</h3>
+      <p>Go create some at {{ source.name }} and come back here to import them.</p>
     </div>
   {% endif %}
 {% endblock content %}
