+++ homebytwo/templates/routes/route.html
@@ -18,7 +18,7 @@
     <h1 class="mrgb0 text-center">{{ route.name }}</h1>
     <h2 class="mrgv0 text-center">{{ route.get_length.km|floatformat:1|intcomma }}km - {{ route.get_totalup.m|floatformat:0|intcomma }}m+ - {{ route.get_totaldown.m|floatformat:0|intcomma }}m-</h2>
     <div class="pull-left mrg-" >&lsaquo; <a href="{% url 'routes:routes' %}">Routes list</a></div>
-    {% if user == route.owner %}
+    {% if user == route.athlete.user %}
       <div class="pull-right mrg-"><a href="{% url 'routes:edit' route.id %}">Edit Route</a> &rsaquo;</div>
     {% endif %}
   </header>
